requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
rembg==2.0.50
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
